Move crawl & parse into their own directories
