+++ packages/word-definitions/src/crawl/index.ts
@@ -1,0 +1,1 @@
+export { default } from './crawl';
