+++ packages/word-definitions/src/parse/index.ts
@@ -1,0 +1,1 @@
+export { default } from './parse';
